Bug 806407 Fix type in struct passed into msgsnd f=jesup
UltraBlame original commit: 891f2dc392f01037dcb48fc50d5d74d4a9c0e709

+++ media/webrtc/signaling/src/sipcc/cpr/linux/cpr_linux_ipc.h
@@ -20,7 +20,7 @@
 
 
 struct msgbuffer {
-    int32_t mtype;    
+    long    mtype;    
     void   *msgPtr;   
     void   *usrPtr;   
 };
